add file share app
