--- fileshare/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('upload/', views.upload_file, name='upload_file'),
-    path('files/', views.file_list, name='file_list'),
-]
